Remove alternative breathing settings and add run_on_startup
