+++ mc/gui/breathing_settings_wt.py
@@ -1,7 +1,5 @@
 import os
-from PyQt5 import QtCore
-from PyQt5 import QtWidgets
-from mc import model, mc_global
+from mc import model
 import mc.gui.toggle_switch_wt
 from mc.gui.reusable_components import *
 from mc.gui.breathing_phrase_list_wt import BreathingPhraseListWt
@@ -125,20 +123,18 @@ class BreathingSettingsWt(QtWidgets.QWidget):
         grid.addWidget(HorizontalLine(), 2, 0)
         grid.addWidget(QtWidgets.QLabel(self.tr("Notification type")), 3, 0)
         grid.addWidget(notification_type_qgb, 4, 0)
-        grid.addLayout(notification_interval_qhl, 5, 0)
-        grid.addWidget(QtWidgets.QLabel(), 6, 0)
-        grid.addWidget(QtWidgets.QLabel(self.tr("Notification Audio")), 7, 0)
-        grid.addLayout(notif_audio_qhl, 8, 0)
+        grid.addWidget(QtWidgets.QLabel(self.tr("Notification Audio")), 6, 0)
+        grid.addLayout(notif_audio_qhl, 7, 0)
+        grid.addWidget(QtWidgets.QLabel(), 8, 0)
         grid.addLayout(dialog_qhl, 9, 0)
         grid.addWidget(HorizontalLine(), 10, 0)
         grid.addWidget(QtWidgets.QLabel(self.tr("Do you always want the same phrase or a random one?")), 11, 0)
         grid.addWidget(dialog_type_qgb, 12, 0)
-        grid.addLayout(show_after_qhl, 13, 0)
-        grid.addWidget(QtWidgets.QLabel(self.tr("Dialog Audio")), 14, 0)
-        grid.addLayout(prep_audio_qhl, 15, 0)
+        grid.addWidget(QtWidgets.QLabel(self.tr("Dialog Audio")), 13, 0)
+        grid.addLayout(prep_audio_qhl, 14, 0)
 
         # second grid column
-        grid.addWidget(self.phrases_qlw, 0, 1, 15, 1)
+        grid.addWidget(self.phrases_qlw, 0, 1, 14, 1)
 
         vbox_l2 = QtWidgets.QVBoxLayout()
         vbox_l2.addWidget(H1(self.tr("Settings for Breathing")))
@@ -187,6 +183,7 @@ class BreathingSettingsWt(QtWidgets.QWidget):
             return
         mc.model.SettingsM.update_breathing_dialog_close_on_hover(i_checked)
 
+    # TODO find out where this was used
     def on_dialog_audio_toggled(self, i_checked: bool):
         if self.updating_gui_bool:
             return
